postgres_etl/requirements.txt
psycopg2-binary>=2.9.5
pandas>=1.5.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
numpy>=1.21.0
watchdog>=2.1.0